==> 2-precedent_preprocess.ipynb <==
import pandas as pd

df = pd.read_csv("data/raw/precedent.csv")

df["판시사항"] = df["판시사항"].apply(lambda x : ''.join(x.split("\r\n")).strip())
df["판결요지"] = df["판결요지"].apply(lambda x : ''.join(x.split("\r\n")).strip())
df["참조조문"] = df["참조조문"].apply(lambda x : ''.join(x.split("\r\n")).strip())

df = df[["판시사항", "판결요지"]]

df["판시사항"] = df["판시사항"].apply(lambda x : None if x=='' else x)
df["판결요지"] = df["판결요지"].apply(lambda x : None if x=='' else x)

df = df[["판시사항", "판결요지"]].dropna().reset_index(drop=True)

df["판시사항"] = df["판시사항"].apply(lambda x : ' '.join(x.split()))
df["판결요지"] = df["판결요지"].apply(lambda x : ' '.join(x.split()))

df.to_csv("data/precedent_preprocess.csv", index=False)

df.head()
